i've changed the requests to xpath. may as well be consistent instead of having a mixture of css and xpath.
in general its working, but there is a strange behavior for the price. it keeps reading the same value, the value for the first entry on each page, for all entries on that page, which obviously not correct. It changes when the spider moves to a new page, and the pattern repeats.

annoyed