I've tried a few things. The first is saving to csv instead of JSON. I'm happy with the result. I like it better than the json.
the second is in agropost_3. I'm trying to pass if one of the fields is empty. I think this has worked as pre-count (today.csv) was 5600 rows. post-count (agr3.csv) is 4800 rows. So there are about 800 rows less.

this raises an interesting question: how best to compare the two and be sure that what I'm excluding is correct?

I have also been exploring the items.py and pipelines. There is a lot of potential and scope but its really hard to understand. I guess it's just a matter of gradually trying, failing and trying again