I've added a little regex to the ```if and``` for date. Based on a quick visual inspection i think its basically OK. there seems to be some odd inconsistency in dates while i was watching it running but i can't locate the records in the ccsv now.
I need to do some more regex for the ```if elif else``` block for number as there are some weird ones in that column.

I need to get the dates in a more consistent format but I'm not convinced that should be done in the scraping stage.

I am still concerned about validation. How can i be sure that what I'm collecting is accurate and correct? I don't know the answer to taht yet