I've made another change to the price part of the parse function. It should now only accept numbers of 3-4 digits and no more than 2 decimal places. This has resulted in a large decrease the total number of rows to 3430. a quick visual inspection indicates that it looks ok but as repeated in many previous commits, i'd like ot have a more objective, quantifiable way to determining the data is indeed complete and correct
it seems to be working.

there are some odd bugs. I notice that the dates around rows 213 - 228 look wrong. It's as if the scraper has ignored the xpaths and split for those lines. There are also some other odd characters that have slipped through. strange.

it is very interesting to watch the feed on the terminal pane as it scrapes. I wonder whether it is possible to record this as a log file like we've done in the past for ML training?