interesting problem appears once i start cleaning the data. I'm glad i changed focus a little.
for the date records that have the weird html problem, it seems that there is also a date problem: "12 Mar Jan 2019" or "20 Feb Jan 2019" and so on. I didn't notice it in the scraper return but i do notice it in once i start trying to clean that data

There is definitely more work to do on the scraper.